Deprecate system::role for DE test cluster
system::role is a bit of a hack. In 21fa33b455bfd956f45525b7bb3c22c77ff3787d
support was added to move the functionality into profile::base::production
and simply set the role description via a Hiera flag.

This converts an initial batch of DE roles in the test cluster.

Change-Id: I769466b4da112a89fcfd6fc5064df8c5335dad79