stewards: Allow lists.wm.o to access the stewards rsync server
When we moved lists.wm.o from lists1001 to lists1004, we moved away from
having service IPs. This turns out to not have been a great decision and
we currently have two PTR records for the same IP. This allows both
answers to access the rsync service until we fix this.

Change-Id: Ic43a49cf3da822653aecff078898facce12ccba0