aphlict: remove duplicate sytemd timer for logrotate
Noticed because sometimes we got alerts for failed unit logrotate.

The failure was due to:

"logrotate does not support parallel execution on the same set of logfiles"

which then leads to:

logrotate.service: Main process exited, code=exited, status=3/NOTIMPLEMENTED

which ultimately causes the failed systemd unit.

The reason for that is there are 2 systemd timers here.

The generic logrotate.service and this logrotate_aphlict.service both
run:

ExecStart=/usr/sbin/logrotate /etc/logrotate.conf

/etc/logrotate.d/ is included there.

The code that drops the aphlict config file into /etc/logrotate.d
should stay and will still be used.

But we should not have another service/timer running that clashes
with the generic one.

Seeing the "TODO: remove once puppet has run" comment here of course
supports the idea that this was simply forgotten.

Assumes we manually delete the unit on aphlict* when merging this.

Then we should never again have new tickets like T367960.

Bug: T367960
Change-Id: Ifb261bc91002addab2b18a5e7efa1707483bf439