Netbox 4: create new script directory and enable debug
Netbox changed its way of managing scripts, just pointing to the local
netbox-extra directory isn't enough. Scripts needs to be imported from a
"Data source".
The datasource can be a local directory or a git repo (or S3 bucket).
And have a button to sync it manually. Automatic sync might be possible.
Once the datasource is created, in the script menu (and script also
include reports now), and admin can either upload a script, or import it
from a datasource. Here we can set it to update automatically when the
datasource is synced. When imported, the script file is copied to the
`SCRIPTS_ROOT` directory.

This patch creates a new directory /srv/netbox/customscripts and sets it
as `SCRIPTS_ROOT` in Netbox's config for `netbox-dev` only.
That way it doesn't interfere with the local `netbox-extra`.

`netbox-extra` is still needed for validators, as the deploy script
creates a symlink from src/netbox/validators towards it.
We could revisit in the longer run on how to make it cleaner.

Hosts: netbox1002.eqiad.wmnet,netbox-dev2003.codfw.wmnet
Bug: T336275
Change-Id: Iedbdadd9e53a448ddb9aec39562eac33347a1f74