Enable the MariaDB binlog on the analytics mariadb replicas
We wish to start experimenting with Change Data Capture tools such as
Paimon, with the intention of streaming changes in the MariaDB databases
into the Data Lake.

The binlogs on the dbstore servers will be a suitable source for this
list of changes, so this commit enables the binlog functionality.

We will retain 7 days' worth of binlogs, with a 1 GB file size of each
log.

Hosts: dbstore1007.eqiad.wmnet,dbstore1008.eqiad.wmnet
Hosts: dbstore1009.eqiad.wmnet
Bug: T370354
Change-Id: I6aa985f83f4cd609d3512f0909202ef189f54bd5