dbbackups: Set sql_mode as loose so that invalid enum values are ok
Some wiki dbs (none critical, mostly non-wiki tables) contain enums
with 0 value ("invalid"/"n/a"), which was not expected, specially
because most of those have a 1 value of "unknown". This is because
they are being imported using a TRADITIONAL / more restricted sql
mode, which leads to import errors.

Respect the original invalid values and set sql_mode temporarilly
as '' (no restrictions/validations) so that imports go through.

Bug: T367162
Change-Id: Ic3944bab0ab4ebc32ea01c495aafb6104c9e185d