wdqs: enable throttling only for reqs from the CDN
Use the header X-Client-IP to identify external requests.
Set the X-Disable-Throttling accordingly if the request is internal.
Configure the throttling servlet filter to be enabled only if
X-BIGDATA-READ-ONLY is set *and* X-Disable-Throttling is *not* set.

Explanation of the nginx map syntax:
if http_x_client_ip is empty then we set internal_request to 1
else set internal_request to empty string
if internal_request is empty str then X-Disable-Throttling not set

Hosts: wdqs1020.eqiad.wmnet
Bug: T361950
Change-Id: Ifb47c7fe6ec3bf5829dd593fe5046fc3420d709a