dynamicproxy: Clarify error page titles
Make the error page titles more specific when possible, and remove the
'Wikimedia Cloud Services' part of the generic error title to make it
more clear the error is most likely with the individual project and not
with the infrastructure. Also, replace the default 'Wikimedia Error'
text in the <title> block with a more specific message.

Also remove an error page file that's not referenced in the
configuration.

Change-Id: I02ab42a97a2d2857b16f2a1a91503e438b0883cf