Update gnmic config to allow processing of all interface stats
This patch removes the queues-counters-only gnmic processor, which
had the effect of only processing interface stats with 'queue' in
the name.  As discussed on task removing this will allow the regular
interface stats which are useful to have in prometheus.

I've also added some new processing rules to not collect stats for
disabled interfaces.  This works by creating a new tag from
/interfaces/interface/state/enabled, then using the event-drop
processor to drop any metrics for matching interfaces.  After
the filtering is done we can drop the 'enabled' tag rather than
store it on each measure, so I've added it in delete-tag-name too.

Tested locally on my system and working as expected.

Bug: T326322
Change-Id: Ie56ef700864b8fa06b806ff33f544ba7efeccfc2