dbbackups: Disable regular es backups (es6, es7) while es4/5 run
Disable bacula backups for External Storage read-write backups
(es6 and es7 ones) while the es4 and es5 ones are being generated
on backup1002 and backup2002.

We will do later a manual run of those at an appropiate time, as
well as of the read only ones for long term storage. Then we will
revert this change.

Bug: T363812
Change-Id: I1dbcb8d283d129c97e27b1f83b7ed277f3fd6de2