gerrit: fix motd/role description
I4fb92833e8cf7f4a5a08829c9b7854180b874b93 moved the role description
from a system::role (in which the $::realm variable is expanded by
Puppet) to a hieradata (which does not expand the variable).

The description is used to forge a script in /etc/update-motd.d which
uses a printf that complains cause it is not given data to expand
${::realm}.

Remove the realm variable from the role description and simply use
"Gerrit Code Review server".

Change-Id: I676c10bdcf3ab94f39e302cdbda1d0481f40a28c