Add ripgrep and fd-find to standard packages
These are Rust-based rewrites with a fresh take over the traditional
GNU utilities. I've been using them personally for quite some time
over the traditional tools and it would be useful to also have them
on production servers. And I saw that Brian recently asked for it
on IRC as well.

The command names don't clash and existing scripts will continue to
use the GNU variants, so there's no risk of regressions and the packages
are really small without external deps.

Change-Id: Ied034eab6dc12750f0a76a93a955945babd91962