Update ceph server firewall and permit access from dse_kubepods
We are at the point of enabling access to the Ceph daemons from the
dse-k8s cluster, so we need to open the firewall on cephosd1*
accordingly.

We take the opportunity to switch from ferm::service to
firewall::service, which will enable compatibility with the nftables
based firewall provider.

Hosts: O:ceph::server
Bug: T327259
Change-Id: I4cab13da68412493fc1f005548011c533367f2d4