profile::puppetserver::git: add an option to exclude servers
In case a puppetserver/puppetmaster fails for a long time,
we need to be able to exclude it from the various configs,
most notably the one used by puppet-merge and private-merge.

For the Puppet 5 infra we have an explicit list of nodes,
meanwhile for Puppet 7 we rely on the puppetservers' role
to get the list.

Bug: T368023
Change-Id: I85065f4b4bfe5319f4c962623dbda44aaec26195