Change gnmi sampling interval and enable timestamps for prom output
This patch changes the frequency which routers export data to gnmic
to 60 seconds.  Reason being promethteus is only scraping the data
every 60 seconds, so the routers pushing it more than that does
not give us grater granularity.

Also enables the 'export-timestamps' option for the gnmic prometheus
exporter.  Which means the timestamp - set to the time the router
sent the data - is exported with each metric.  This removes the
current inaccuracy whereby prometheus is using the time it scrapes
the data from our netflow hosts when putting it into the db.

Lastly increases the scrape_timeout from the default (10s) to 20s as
it seems to be taking just over the 10s in codfw right now.

Bug: T326322
Change-Id: Id0edadb588b7b0b35d9c20819134c941c3f9c9e7