hiera: dns6001: reduce anycast_hc logging level and backups
By default, anycast-healthchecker sets the log verbosity to 'info'
(log_level) and maintains 8 backups of log files (log_backups). While
this frequency of logging is helpful initially, there is no reason for
us to log every successful healthcheck.

In I9f81e79577, we added support for customizing the log levels and the
number of backups we should maintain, so set that for dns6001 to start
with. If this helps reduce the CPU anycast-healthchecker consumes as
well (which seems to be more than gdnsd itself), we should roll this out
to everywhere and delete the old logs.

Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: I9fbd5f7356a5011dcde3ab852b9071bbe2fa4ec0