logstash: add normalize_labels script
ECS labels should be a string, but this is not enforcable by the template.

OpenSearch guesses the label type and indexes it as keyword, but this can
still result in a race condition where the first label encountered as one type
causes all other events with labels as any other type throw an
illegal_argument_exception and become dead-lettered.

By normalizing on string for labels, we will eliminate this vector for
dead-letters.

Bug: T368867
Bug: T342451
Change-Id: Ifb5325969b169c634de19f567418b54e64a3dfe9