benthos:cache: encode problematic fields as b64url
To avoid a probable bug where HAProxy doesn't send valid UTF-8 in
RFC5424 syslog messages (and therefore Benthos fails to parse it), we
decided to encode referer and user agent fields (other fields can be
encoded too, if needed) and decode them later in Benthos.

The "base64rawurl" encoder isn't listed in the official Benthos
documentation but it's the one that works better in both padded and
not padded outputs from haproxy.

This broke unit-tests that needs to be completely rewritten to support
encoding on all headers.

Bug: T365718
Change-Id: I4bf4ea8850cf3a93ba043bf093c7e7aca1b0d777