P:kubernetes:node: Autorestart ferm.service
Why:
ferm.service restart by puppet and kubeproxy changes to iptables enter a
race condition leading to bad iptables status and connection issues
between services, in particular eventgate. This results in an exit code
of 4 for ferm.service.

What:
Introduce a systemd::override to autorestart ferm.service when it exits
in a failure state. It will try to restart every 5s, with a limit of 10
times in 2 minutes.

Hosts: auto
Bug: T354855
Change-Id: I1bb773cdfd6bfa67668b706fab235a6346e8f5e1