puppetmaster: change git sender email address to git@wikimedia.org
T369004 describes why it would be better to use a real address
as the sender address for git mail about commits in the private repo.

My experience is we can use any @wikimedia.org address as the sender
and this is an actual Google account:

[mx1001:~] $ sudo exim4 -bt git@wikimedia.org
git@wikimedia.org
  router = gsuite_account, transport = remote_smtp

It also fits the purpose, though I am not sure what
exactly it was created for in Google.

Change-Id: I828e0e54a6e7c390c612d825e75b9d85f1d73363