phabricator: add script to set safedir git config for deploy repo
We need to set git safedir but only for the actual Phabricator
deploy repo.

The deploy repo path is 2 symlinks removed from /srv/phab
and changes on every scap deploy.

We don't want to just set safedir systemwide on '*'.

So we have to do some scripting to get that properly done
since we can't get that with just puppet code alone.

If we don't set this we have side-effects such as not displaying
version info which this is fixing.

Bug: T360756
Change-Id: Ibf55af0d8ef8f607b06b31bc66ee6aaecc670781