deployment::rsync: Add support for PKI
Why:
deployment::rsync forcefully sets up stunnel for syncing the various
components between the deployment hosts. This is correct and follows
best practices. However:

* deploy1002 and deploy2002 use Puppet 5 and puppet 5's CA
* deploy1003 has been switched to Puppet 7 and has the new CA
* stunnel uses the Puppet PKI certs for everything and verifies certs
  and CAs
* stunnel on deploy1003 doesn't trust stunnel on deploy1002

What:
Request a certificate from our PKI infrastructure. The certificate is
still gonna be FQDN based, tied to the host, but not couple of our
Puppet CA and Puppet versions.

Alter the various ERB script and puppet classes involved to use the
paths of the certificate, key and PKI CA instead of the Puppet CA ones

Hosts: auto
Bug: T364417
Change-Id: I201563a0d1f02cfde2e035bdf91f6b2a5b30b215