merge_cli: fix a puppet-merge.sh comment
The note about the ssh's configured command should be placed
on top of "merge()", as it was in the past before the code
was changed/moved around.

Add also a note about where to find the ssh forced command
so it is easier to get how it works in details (this particular
bit of code is very important and essential in my opinion).

Bug: T366355
Change-Id: I845b62f1dc23fe80c2fb1b2f40c7f6a62d7ba96b