puppetmaster::gitclone: disarm pre-commit and post-commit hooks
We are transitioning away from puppetmaster1001's /srv/private
repo in favor of the one on puppetserver1001. In order to do it,
we should disarm the pre-commit and post-commit hooks, warning
the user about the usage/state of the repository. This is only
a transitional step to ease the migration, and have a quick way
to rollback in case something doesn't work.

The idea is to make this repository almost like a bare one, but
without really changing it (so the rollback is quicker if needed).

The change should be merged only after the move is properly tested
and announced.

Bug: T368023
Change-Id: I6d3baaf616f4a6089b5366618753869a783d7db5