Add the analytics-wmde keytab to the stat boxes
In order for members of the analytics-wmde-users group to be able to
interact with kerberised services as the analytics-wmde system user, we
need the analytics-wmde keytab to be present on the stat servers.

This is in order to support commands such as `yarn logs` for
troubleshooting purposes.

Hosts: O:statistics::explorer
Bug: T340648
Change-Id: I3081c6cd83401b18573ec9d6153c3b9539f82b1a