Disable monitoring on clouddb1021 prior to decommissioning
We have migrated the role previously performed by clouddb1021 to a new
host an-redacteddb1001. In order to help ensure that we have completed
the migration successfully, we plan to disable the MariaDB services on
clouddb1021 for a week or so, prior to its decommissioning.

This patch disables monitoring on clouddb1021 so that we can stop the
services without generating alerts.

Hosts: clouddb1021.eqiad.wmnet
Bug: T368518
Change-Id: I23334d5a111f2ba75994fbd1c93279361c61fcf2