memcached: enable extstore to eqiad only
Why:
Throught discussions we came to the conclution that
in order to properly assess if extstore is serving us
well, we will have to do an A/B like test between eqiad
and codfw

What:
We are enabling extstore on all eqiad servers and completely
disabling on codfw.

This change requires respected servers (~all eqiad + few codfw)
to have the service restarted. Thus we will disable puppet
on all mc* hosts, and gradually enable it as we are slowly
restarted.

Hosts: mc2050.codfw.wmnet,mc1039.eqiad.wmnet,mc1050.eqiad.wmnet
Bug: T352885
Change-Id: I5bc936a07f3e1ce93bed16c9781fa7772b970694