ats: Route /api/ to /w/rest.php on mw-api-ext
Why:
The thinking in layed out in T364400, but here's a summary. Note that
there are 2 different concerns here:

* One concern is the overall theme layed out in T364400
* The other one is which mw backend release will be serving /api/, as
  for separation of concerns reasons, we want end user and API traffic
  separate

For the former, we want to consolidate the 2 different APIs into one,
this is a first step, there's more to be figured out.

What:
* Augment the regex matching /w/rest.php to also match /api/ and direct
  traffic for /api/ to mw-api-ext

Note that this doesn't touch on T364400#9780622 et al

Hosts: auto
Bug: T364400
Change-Id: I68c7a5b0890d5d4185e56c24fdc0218ff4146541