Puppetize the disabling of the misc dumps on snapshot1017
We have disabled the dumps on snapshot1017 manually, but we now wish to
re-enable puppet, since it has been disabled for too long.

There are no useful hiera switches to enable/disable the dumps, so for
now we can simply omit the profile that creates the resources.

We will revert this change in the coming days, as we proceed to
re-enable the misc dumps.

Hosts: snapshot1017.eqiad.wmnet
Bug: T368098
Change-Id: Ia407ee985b1c00f6b54f520340d9de892b23ae53