Configure analytics_meta MariaDB clients to connect to an-mariadb1002
We are performing a role swap between an-mariadb1001 and an-mariadb1002
in order to support an upgrade of the switch to which an-mariadb1001 is
connected.

This will require a coordinated reconfiguration of the hive metastore
services and the druid brokers, whilst manually updating the replication
settings of the MariaDB servers.

Hosts: O:analytics_cluster::coordinator,O:analytics_cluster::mariadb
Hosts: O:druid::public::worker,O:druid::analytics::worker
Bug: T365503
Bug: T348977
Change-Id: I83e0437837411332094ffe674b49b042447dca5f