Temporarily pause gobblin to facilitate Hive/MariaDB maintenance
We need to change the active MariaDB server that serves analytics_meta,
which will require a short period of unavailability for the Hive
metastore.

This commit will be merged around an hour ahead of time. It will
temporarily absent the gobblin timers and thereby reduce any attempted
writes to Hive, since the pipelines will simply be waiting for data.

The changes will be reverted after the MariaDB servers have swappoed
roles and the Hive metastores restarted.

Hosts: an-launcher1002.eqiad.wmnet
Bug: T365503
Change-Id: I592b51e81ab0de954699dc6105e2216683a07ca0