admin: add Tyler Cipriani as group approver for contint-docker
As part of T276465 we are trying to add approvers to all groups
that don't have group approvers yet.

This group is applied on releases and CI hosts.

Members of the release engineering team are automatically members
of this group but a few other users are in it additionally.

As of today:

members: [*ops_members, *releng_members, gjg, addshore, reedy, krinkle, jforrester]

Since Tyler is already the approver for releng and other releng-related
groups suggesting to also make him the approver here.

Bug: T276465
Change-Id: Iaa7ee9c26eb47d0183e10d96661ff58bc0c42097