gerrit: switch gerrit-replica from iptables to nftables
This is part of a general move from iptables to nftables
(see tasks linked to T348498 etc).

But more specifically we want it here to be able use features
that were only in the "recent" module/plugin for iptables
but come per default with nftables.

First only doing this on the replica host and not the main
server. Later moving it to role-level.

Change-Id: I21a9c6e64cec70d98f02e845d28df95d865fedbf