admin: add Kwaku as approver for the dns-admins group
As part of T276465 we are trying to add approvers to all groups
that don't have group approvers yet.

This group is for "anyone who should be able to merge DNS changes
(and doesn't already have global root).

For right now this is used only for fr-tech admins but theoretically
could be used for others in the future.

Per IRC chat the traffic team should own DNS so suggesting to add
Kwaku as the traffic team manager as approver.

Bug: T276465
Change-Id: I06e2f4ce94900be9411192b164d79e138ba79970