crm: Add lookup for civicrm branch to check out
We want the ability to have a testing branch that is applied to the labs
hosts. This allows us to use the same puppet config and change the
branch in the labs world independantly.

It also shifts the git clone to using checkout. Since we want to force
it to the specified branch so we need to use checkout instead of latest
for the clone.

It should hopefully allow us to get rid of the standalone puppetserver.

Also, some typo fixes.

Bug: T343486
Change-Id: I9b70321c4aa0857cbac26754b813fb0c816e023c