redirects.dat: change funnel target for sep11.wikipedia.org to meta wiki
Historic wiki that was then deleted and back in 2015 replaced with
links to the Wayback Machine (I4e2af4454ea9d4).

But that links to a list of snapshots and not even a specific version.

Per T367014#9971812 linking it to https://meta.wikimedia.org/wiki/Sep11wiki
would actually explain the history of this URL.

If we change this once to a wiki, and meta seems just right for that,
then links to Wayback Machine can be added to the wiki page as needed.

Bug: T367014
Change-Id: Ic94f3b3bddf1d360facdac41e3596000a897ab58