profile::kerberos::kadminserver: allow more nodes in rsync
This is a step needed to be able to transfer the keytabs to
puppetserver1001 as well. The puppetmaster1001's usage is going
to fade away very soon.

It seems ok for the moment to just allow one puppetserver, rather
than all of them, but we can review the choice later on if needed.

Bug: T368023
Change-Id: I2919e48e7b8b7b89416d3f3a2b6cd4b4a8029c6e