Add the thirdparty/yarn deb repository to reprepro
We need to be able to install the yarn package manager in our nodejs
environments in order to create a container image for growthbook.

This pre-2.0 version of yarn is considered legacy by the upstream
project and installation from packages is to be replaced with corepack,
but this version would still be useful to us.

Bug: T365839
Change-Id: I5e0a98edf37164f4229a45a3511fb4c0cdf4be4c