gitlab: switch gitlab from iptables to nftables
This is part of a general move from iptables to nftables
(see tasks linked to T348498 etc).

But more specifically we want it here to be able use features
that were only in the "recent" module/plugin for iptables
but come per default with nftables.

Switching the replica only is not possible because of the required
puppet code refactoring.

Bug: T366882
Bug: T365259
Change-Id: I55a4d8e658c60108deae61b8ed84210c3b29c395