cfssl: add a condition to cfssl_ocsprefresh.py
For some reason in cloud we don't always have a last update time
on the DB for ocsp_responses (maybe due to the different and limited
setup). Most of the ocsp refresher jobs are failing since a long time
due to NULL/None returned from the DB. Add the condition to the Python
script to fail gracefully.

Bug: T363829
Change-Id: I79d881bd0dae2474b22a8dedc0e37dbb13582115