profile::tcpircbot: allow puppetservers to contact tcpircbot
The new nodes will replace the puppetmasterXXXX ones very soon,
so we should allow them to talk with tcpircbot to have logs
from confctl properly logged.

Note: I added the nodes only to the tcpircbot::instace's cidr
parameter before this change, and of course nothing happened.
Not sure if the cidr parameter is still needed or not.

Bug: T368023
Change-Id: Ifef34a31296efdc677e933d8223065d206761f94