Disable produce_canary_events systemd timer
This job has been migrated to Airflow.  We've kept it running as a systemd
timer for a while in parallel to evaluate the Airflow job's effectiveness.
The Airflow job has been reliable for a while now.

We need a later version of refinery-job in order to unblock SRE's migration
of MW to MW on k8s.  refinery-job 0.2.44 does this, but it also has the latest
ProduceCanaryEvents job, which no longer works for systemd.

To unblock SRE, we will disable the produce_canary_events systemd timer.

Bug: T370186
Bug: T367949
Change-Id: Ia85cd5abfeb3ff98813db91d99d117bdad3026ab