admin: add dcops to the system adm POSIX group
The DCOps team asked to be able to connect to prod
nodes and collect basic info about a system.
The motivations are multiple:
- Learning experience for the team members.
- Less pain while gathering basic debug info about
  a failed node.

The proposal is to:
- Remove sre-admins from datacenter-ops to better
  capture the team's members.
- Create a new group called "ops-limited", to
  replace sre-admins, since we hope that a better
  naming will reflect the real purpose of the group.
- Add dcops folks to the ops-limited group, that
  should allow them to get ssh access to nodes in
  production, with some extra sudo privilege
  (mostly for debugging purposes).
- Add ops-limited to the 'adm' POSIX group, so
  members of that group can read files like syslog,
  httpd logs, etc..
- Move user tappof from sre-admins to ops-limited.

Anything that can modify the status of a production
node (raid array, ip config, etc..) should require
a sync with an SRE belonging to the group that owns
the target host (see also efforts for sudo_pair in
T299989).

This change will be followed up by a complete removal
of sre-admins.

Bug: T360356
Change-Id: I535868b134b5a3d67d107e9420f6ffb4d676f5f6