labs_lvm: fix volume creation using relative size
When creating an instance with a relative size such as `100%FREE` we
went to pass lvcreate "-l 100%FREE" as a single argument. It confuses
the command which fails:

+ /sbin/lvcreate --wipesignatures n '-l 100%FREE' -n second-local-disk vd
  Invalid argument for --extents:  100%FREE
  Error during parsing of command line.
+ exit 1

That was caused by passing the script through shellcheck in a2f7e3cf51.
Although in other files and instances I have properly quoted separately
the option and its value, in that specific case they were quoted
together leading to the above issue.

Bug: T370312
Change-Id: I2125e00d52c21de1ddab211bdbdb1c563f502a34