pcc: improve error checking
Prior to this change if a host existed in puppetdb, but only had facts,
pcc_facts_processor would not compile a catalog for the node. This meant
if compilation failed, but facts upload succeeded, we would never try to
compile a catalog for the host.

In addition we tried to compile a catalog if any error was returned from
puppetdb, however, we specifically want to check for a 404, other errors
should be raised.

Bug: T367547
Change-Id: I41c7e8eb298f2a57ca19c79bb04991ef37e80ad5