expose_agent_certs: use ssldir exclusively
Prior to this commit we used a mix of ssldir and $facts['puppet_config']
to determine the location of tls material. The latter method is more
robust to future config changes, but breaks users of this define type,
such as pcc puppetdb, which override the ssldir param. After this commit
we use ssldir exclusively, if we or puppet changes the location of tls
material in the future, we will burn that bridge when we get there.

Hosts: auto
Bug: T367547
Change-Id: I210f74b8d2f94a4a47eb78c202a9de21fd9e3b3e