mediawiki-cache-warmup: support 'clone' for mw-on-k8s
Basic idea: use the kubernetes API to enumerate pod IP : port pairs
backing the endpoints object associated with the inbound TLS service.

Since pod IPs are routable, this totally works.

To support this, the host where one invokes warmup.py changes from a
maintenance host a deployment host, and commensurately the switchdc
00-optional-warmup-caches.py cookbook (and spicerack.mediawiki) are
updated provide this functionality (separate patches).

In addition, the "dry" subcommand is removed, in favor of showing a
sample of request URLs and targets by default, and exiting without
sending requests if the `--dry-run` global option is specified.

If the full URL and target lists are desired, use the `--full` option.

Bug: T369921
Change-Id: Ic48417e5acb0a64cd6af1c66a2b25853a8c2a5ef