Increase the heap for the mapreduce history service on an-master1003
We have seen that the mapreduce history service has been crashing with
oom errors lately, but only in production. This commit causes a new file
called /etc/hadoop/conf/mapred-env.sh to be deployed on an-master1003,
containing an environment variable that will configure the heap size on
the MapReduce history server to be 2048 MB instead of the default 1000 MB.

It does not apply the same change to the analytics-test-hadoop cluster,
since this is not affected by a memory shortage.

Hosts: an-master1003.eqiad.wmnet, an-test-master1001.eqiad.wmnet
Hosts: P:hadoop::common
Bug: T369278
Change-Id: Ia8fdf6b16d705d6f6e7747c73ce1d7efb10c0c6e