kafka: use instance-based selection for kafka-kit storage metrics
In the kafka-logging case the brokers are co-hosted in the same cluster
as other hosts. Switch the Prometheus query to look for each broker
disk space explicitly, to make prometheus-metricsfetcher work out of the
box.

Whereas kafka_log_size can stay with 'cluster' label selector since the
metric is already kafka-specific.

Bug: T370129
Change-Id: Id34fdeb27ef44ea528cff396b7a0d08883018d41