Netbox 4: add version flag
To be able to run Netbox 4 and Netbox 3 fully puppetized in parallel.

Also Netbox 4 report timers: run as sre_bot user
If not specified in Netbox 4, the scripts/reports are ran as "gehel"
user. Here we explicitely instruct it to use the sre_bot user as it's
made for that.

Hosts: netbox1002.eqiad.wmnet
Hosts: netbox1003.eqiad.wmnet
Hosts: netbox-dev2003.codfw.wmnet
Bug: T336275
Change-Id: I62630dd000289d276f179c76683f9567fa6310ad