ncredir: Reformat/sort the redirects file
This prepares the nc_redirects.dat file for automatic patch proposal via
ncmonitor. There were many sections in this file but they didn't really
serve much in the way of organization. ncmonitor will automatically
append/sort.

One downside to this is that ncmonitor's handling is pretty naive: It
will mangle any and all comments that are not part of the top clump of
comments. A weak attempt to address that, a comment on the top of the
file, will hopefully be enough (commits are not automatically merged but
are approved by a human).

As a bonus, a few spaces were replaced by tabs. Can you find them all
and win today's prize?

Bug: T355189
Change-Id: I9f6e91fe2c11e053ff175b882ab17e5b31fc269b