puppetmaster: set git::clone umask to match requested file mode
git::clone executes git commands using with a default umask of 022 which
matches the usual file mode of 0755.

On the Puppet master, there are two clones of operations/private each
using different file modes: 0700 and 0750 which is to prevent access
from group and/or others.  But git::clone defaults to 0755 which is
inconsistent.

git::clone requested mode is applied to the directory holding the clone
which is sufficient to prevent accesses from the mode.

This will ease the removal of the umask parameter from git::clone which
I have proposed with I6e72813e2e6a637c20a3bd5a455665ea93450fc1

Bug: T338277
Hosts: C:puppetmaster::gitclone
Change-Id: I84ca19faf8c0671ea0ba5dd2bc130bb37fe68e01