phorge: add UnsafeAllow3F rewrite flag
This module is not used in production.

But we adjust it to match a recent change in the phabricator module.

It has been determined that in this case it is not actually a security
issue but if we don't allow apache to rewrite URLs with question marks
certain forms links become 403.

Bug: T370110
Change-Id: I7eff0da122711982518a8130a478649ca62baa97