profile::haproxy: move tls_terminator.pp to profile module
The define for haproxy::tls_terminator was used only once in the code,
and it needed 39 parameters to be passed to it. I moved the code to the
profile as it seemed to be more readable and at the same time it reduces
the size of the generated catalog quite a bit.

Change-Id: Ief3225c3612dfef50afdffaa32d7acdfdae1506b