services_proxy: add idle_timeout parameter, enable in shellbox-video
Currently we are seeing timeouts when connecting from videoscalers to
shellbox-video via the service proxy. The envoy logs report the code
"SI" which explicitly indicates "Stream idle timeout in addition to
408 or 504 response code.". This aligns with the fact that the related
request time is always a few ms off of 5 minutes - the default stream
idle time in Envoy.

Allow the setting of the idle_timeout on a route level for requests to
services via the service proxy where necessary.

Bug: T356241
Change-Id: I85d3617e45c40a73437283c7e5ad049b2e2e5571