wikistats: mount a cinder volume to store backups externally
Created a cinder volume for the project per:

https://wikitech.wikimedia.org/wiki/Help:Adding_disk_space_to_Cloud_VPS_instances#Quick_start

Called it "backup" and attached it to wikistats-bookworm instance
in Horizon.

Copied this puppet code from profile::labs::cindermount::srv
to mount it automatically but at a custom mount point.

If it works we can then add a job to copy files to it and finally
have an external DB backup that lives outside the current instance
and can be mounted on a new instance in the future.

Change-Id: Id2f6a83092722d5d427a127a8701f5f2468c8a57