profile::docker::reporter::report: add min_debian_version arg
With the new version of docker-report we can specify a minumum
debian version to support, that is inspected for every image
at runtime before running debmonitor on them. The current default is
10, but we cannot support it anymore since we cannot install
debmonitor on Buster.

The images not supported or discarded are logged by docker-reporter,
so we'll have a way to know which ones needs to be dropped/deprecated.

Bug: T367427
Change-Id: If04deca2db959e79a06ec39712e909e0ec956a55