cumin: set git::clone umask to match requested file mode
git::clone executes git commands with a default umask of 022 which
matches the usual file mode of 0755.

On Cumin, the Homer public and private repositories are cloned with mode
0440 which is to prevent access from others but git::clone uses the
default umask of 022 which is inconsistent.

Set the umask to 337 which matches the requested file mode of 0440.

This is similar to the change I have made to the clones on the Puppet
masters I84ca19faf8c0671ea0ba5dd2bc130bb37fe68e01.

This will ease the removal of the umask parameter from git::clone which
I have proposed with I6e72813e2e6a637c20a3bd5a455665ea93450fc1

Bug: T338277
Hosts: C:homer
Change-Id: Id55dec816ddcda3f3a53434d0eef95d34f4ee7cc