gerrit: use list of replicas from hiera again, don't do puppet DB lookup
This code, which looks up a list of hosts using the gerrit role in
puppetdb, can create a chicken and egg problem as mentioned in T367547#10015192.

In a situation where we get a new puppetdb that doesn't have hosts yet
suddenly we can't compile changes anymore and every puppet run fails
on the compiler machines.

Only after a succesful puppet run hosts can be in the database but
without any hosts in the database we can't have a succesful run.

Since we still have the list of hosts in Hiera that we used before
I suggest simply going back to using that.

This removes that entire dependency on a working puppetdb and simplifies
things while the price for it is just adding a new host to a yaml file
about once every 2 years.

Additionally when we get new hosts we might also like having the flexibility
to be able to apply the role _before_ declaring a host an official replica.

Change-Id: I8bc3cda09f244a06928b21d7e1965e7d9adc4861