cinderutils: allow floating point numbers for min_gb and max_gb
Currently running into a problem trying to use this code to mount
a cinder volume of 1GB size.

If I set min_gb to 1 then it's still not considered "larger than 1GB"
and puppet fails to mount it.

If I set min_gb to 0.8 then puppet fails because it only accepts
integers.

If I don't set min_gb at all then it defaults to 10GB and won't mount
it because it's not larger than 10GB.

Looking at the calculation further down where it's used it seems like
using floating point numbers shouldn't be a problem.

So that would let me set it to 0.9 or so.

Bug: T371573
Change-Id: Id4682b178c8fb2ce95151c61862306cf2eb5fc9c