analytics.wikimedia.org: improve caching and redirects
Caching
  Some confusion arose from the cache for the files not matching the
cache for the HTML file browser interface generated by mod_autoindex, as
a solution here we extend the same caching to all paths where we serve
files, namely analytics.wikimedia.org/published...

Redirects
  There used to be a file server hosted at datasets.wikimedia.org.  When
analytics.wikimedia.org was set up, we migrated the files from datasets
and put them in the published/datasets/archive folder.  The users and
use cases of those files are long gone, so we can change this redirect
rule to just point to the root of all files, as this has caused some
confusion too.

Change-Id: I23c41dfdc0f70a4bd5dc92eb9a8940669170eb5f