planet: disable wikimedia.pt feed causing update errors
This feed was recently added as part of Iad07a4afa5ad.

Since then we are (at some times?) getting errors when
updating the Portuguese feeds.

This results in failed systemd units which then create
tickets like T371064.

The parser error given is:

"TypeError: '<' not supported between instances of 'Feed' and 'Feed'"

when pulling this feed. (systemctl status planet-update-pt)

Change-Id: Ie14a993ee4f0490d64f7e3eb2602308cde927c86