gerrit: add parameter for lfs sync destination host name
follow-up to I8bc3cda09f244a0

to fix T367547#10015192

We are currently using gerrit_replica_hosts for 2 different things.

One is the apache virtual hosts, where we want to use the service
name, such as gerrit-replica.wikimedia.org.

The other is the destination of the lfs data syncing, where we
need to use the actual host name, such as gerrit2002.wikimedia.org.

Previously we got this host name from the puppet DB but with that
we rely on a puppet db always working and finding hosts, which
it only does after at least one succesful puppet run.

This separates the 2 things and doesn't rely on puppet db anymore,
for the small cost of having to update this host name once
every couple years.

Bug: T257741
Bug: T367547
Change-Id: Ib02a24cb39f29d6f35658fa7edfdeac1469354e3