git-sync-upstream: execute the entire script as gitpuppet
This partially reverts a2d86848f1aeb4b24c2ad2f27ac97642dbd54805, which
added the ability to run the git operations of git-sync-upstream as a
different user than root. There were concerns in the origial commit that
the methodology of seteuid would not be completely effective and that
seems to be the case as WMCS is still having issues with git ownership.

Instead add the git user to the prometheues-node-exporter group, so it
can write to the exporter textfile directory.

Bug: T364492
Hosts: cloudinfra-cloudvps-puppetserver-1.cloudinfra.eqiad1.wikimedia.cloud
Hosts: puppetserver1001.eqiad.wmnet
Change-Id: I863a71fc57ec8597e6d7df9235da5ad9b002761c