ci: add new ECDSA ssh key for jenkins to connect to itself
The goal here is to replace the old RSA key created back in 2012
that is used by jenkins to connect to itself.

Created a new ECDSA key.

Added it next to the existing key as suggested by Hashar in
T177826#10034509. We expect it to accept them as a single
string added to /etc/ssh/userkeys/jenkins-agent.

The private part of the key will be added to the Jenkins
credentials store.

There is a passphrase on the key that will be shared privately
to add it there. (possible per
https://www.jenkins.io/doc/book/using/using-agents/)

After fighting once again with array vs string/puppet-lint/wrapping long lines/
ugly HEREDOC/etc I got reminded that SSH keys are data and not
code and belong in Hiera anyways. And Hiera doesn't mind long lines.

So moved them there.

Bug: T177826
Change-Id: If89cc73e93054c8ebbd7040afa1834ab7fb8de46