P:dns::auth::update: maintain admin_state via confd
This commit introduces a test to maintain the state of the admin_state
file via confd instead of Git. See T369366 for more information on the
intention behind this change.

To start with as a test, we introduce a dummy confctl_admin_state file,
applied only to dns4004. The idea is to test everything related to this
change but without actually modifying the admin_state file. Once this
change is tested, we will remove this dummy file and switch to
admin_state managed by confd on all hosts.

Bug: T369366
Change-Id: I310affc866062f3763fb3e378d763332e36a0f33
Hosts: dns1004.wikimedia.org, dns4004.wikimedia.org